backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
PyJWT==2.10.1
psycopg2-binary==2.9.9
python-multipart==0.0.6
python-dotenv==1.0.0
requests==2.31.0
openai==1.3.7
google-auth==2.23.4
google-cloud-secret-manager==2.16.4
google-auth-httplib2==0.2.0
google-api-python-client==2.108.0
google-generativeai==0.3.2
pydantic==2.5.0
pandas==2.1.4
numpy==1.25.2